Replaceв double type to decimal as better type for represent money, all unit test were fixed and return success

+++ SaleTerminalLibrary/Interfaces/IPrice.cs
@@ -5,11 +5,11 @@ namespace Epam.Demo.SaleTerminalLibrary.Interfaces
     /// <summary>
     /// Interface that describe price of product
     /// </summary>
-    public interface IPrice : IComparable<double>
+    public interface IPrice : IComparable<decimal>
     {
         /// <summary>
         /// Property value of price
         /// </summary>
-        double Value { get; set; }
+        decimal Value { get; set; }
     }
 }
